pecl-gearman based server for poschp
this is a largish commit containing everything i need to reply to
a simple osc ping message. it is based on pecl-gearman and it
depends on brian moons GearmanManager heavily.

The code works as such (run deamon with ./bin/poschp start) but it
still needs work. I'll try and update the doc more fully for this
to represent the current state of work soonish.